calendar-screen is completely moduelarized and ready to be used in other screens
Needs better styling
